Fix tests and other minor changes (#2)
* Add setter for ImageParams

* revert back to unique pointers for managing the memory in OSEM

* Add Variable.cpp file

* change numDoi to numDOI

* Change environment variable name

* Pre-allocate for batch loading and other sync-related changes

* Add newline to .gitignore

* Remove S

* Add warning for environment variable names

* Fix some errors in the reconstruction script

* Adjust example script

* Rename parameters for BinIterator constructor

* Change tolerance

* Restore image params values to double to restore tests integrity

* Restore usage of double instead of float for Siddon tests

* Adjust example scripts

* Deallocate sensitivity image copy after OSEM reconstruction

* Add safety checks and rename resample function

* Add clang-format call to BinIterator.*pp

* Add include for std::min/max

* Apply clang-format

* Fix comment

---------

+++ yrt-pet/unit_tests/recon/test_Siddon.cpp
@@ -42,12 +42,12 @@ TEST_CASE("Siddon-simple", "[siddon]")
 	std::string rseed_str = "random_seed=" + std::to_string(random_seed);
 
 	// Setup image
-	size_t nx = 5;
-	size_t ny = 5;
-	size_t nz = 6;
-	float sx = 1.1;
-	float sy = 1.1;
-	float sz = 1.2;
+	int nx = 5;
+	int ny = 5;
+	int nz = 6;
+	double sx = 1.1;
+	double sy = 1.1;
+	double sz = 1.2;
 	double ox = 0.0;
 	double oy = 0.0;
 	double oz = 0.0;
